requirements.txt
kafka-python
grpcio-tools
grpcio